requirements.txt
matplotlib
nltk
numpy
pillow
argparse